[Refactoring ⚙️] 뷰어 컴포넌트 분리
